Update html-and-js
This has more options for vests. Keep in mind that the numbers change all of the time so we need to build the script to call the api to get vests and then do the formula to output to the http post. This script works but has to be manually updated from time to time. Need to finish this. Or if someone wants to finish that would be great.